refactor: split fallback_streams.py into 5 cohesive sub-modules (iter30)
Decompose the fallback_streams.py god-file (~2260 NLOC -> 251) into 5 sibling modules
each <500 NLOC: fallback_streams_probe (HTTP probe / manifest cache / fingerprint),
_identity (title/release identity + similarity tiers), _match (peer/profile matching),
_select (streaming selection + public attach entry points), _attach (ranking / pubdate
dedup / job builders). fallback_streams.py is now a facade keeping the redirect-refusing
urlopen machinery, shared literals, and re-exporting every moved name via __all__.

Proven net-positive pattern: moved functions reach module-globals (own + parent) via
function-local `import resources.lib.fallback_streams as _fs` (call-time, so @patch
resolves and no top-level cycle); no params threaded (params>8 flat at 16). Every
cross-module import (resolver/router/stream_proxy) and test-@patched name preserved;
follow-up-fetch dispatch timing + candidate validation unchanged.

Codacy complexity 26 -> 25 (file-nloc>500: 3->2). Verified: 2138/2140 tests pass
(test_fallback_streams 140 incl. timing test; 2 unrelated pre-existing timing flakes),
pylint 10/10, ruff+black clean, vermin 3.8 OK.
